requirements.txt
flask
python-dotenv
tavily
google-genai